Move yield/await to irbuild.statement
